Social media links
These are added as an array too, to support future back end. The links reference to the array itself, so everything can be updated through api.

Didn't add the iconography yet as i don't have the needed dependencies for it yet - will focus on the UI side later. I did however add the property for icon to the array already