Contact form
Added a contact form component and an additional component for showing the form after you'd submitted it.

It is faked for now, as i don't have anywhere to submit it.  This won't likely be a part of the end product, but is required if i want to use it as part of a React development course which i might submit it for.